requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
openai==1.51.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
spacy